I have integrated the bank database to Plutus. The user information and account threshold in the database can be updated.  For the purpose of Plutus development, I have added a backdoor in the login menu, so we can directly access to the Bank database and perform debug and maintainance. Currently, our database simulation can serve and test three activities: simulation, update and retrieval.
Once the bank database is simulated, it stayed in the private storage of app
until the deletion of app. I have added the bank database manager as private
members in each activity, so each activity can access to the database as in read
mode. There are two activities having the both read and write mode:
UpdateAccountSettingActivity and BankDatabaseSimulationActivity. When an
activity is on pause, we close the connection to the database. When an activity
is on resume, we open the connection as either read of read-write mode
accordingly.