Bug Fixes + Orientation
Fixed a bug where the user selecting checking first would cause the app
to crash. Fixed a bug where an orientation change would cause the app to
crash. Also, enclosed all the layouts in scroll views so that if the
content is off the screen the user can scroll through it. Also handled
keeping track of the current layout during orientation change.